apply: fix SVG preview escape causing terminal breakout and RCE
Sanitize syntax-highlighted preview output by truncating long lines
to viewport width before highlighting and stripping non-SGR escape
sequences. Extract shared strip_non_sgr_escape_sequences to a
dedicated sanitize module used by both chafa and syntax highlight paths.
