+++ .opencode/agents/openspec.md
@@ -33,18 +33,16 @@ Steps defined in the flowchart are detailed below if necessary. Loop through the
 
 ```mermaid
 flowchart TD
-    A[Ask user]
-    A -->|User answers via text input| B[Refine current step with additional information]
-    B -->|Refinement finished| E[Commit changes]
-    A -->|User answers with command| C[Call command]
-    C -->|Command called| D[Ask open questions]
+    A[User input]
+    A -->|User inputs instructions| B[Follow instructions]
+    B -->|Instructions followed| D[Ask open questions]
     D -->|Questions answered and artifacts refined| E[Commit changes]
     E -->|Changes commited| A
 ```
 
-### A\[Ask user\]
+### B\[Follow instructions\]
 
-Ask the user what command to run, or what additional information the user wants to provide as text input. The user can choose from the following commands or provide free-form text input:
+If the user inputs a command like `/opsx-<action>`, where `<action>` is one of the following, execute the command:
 
 - `ff`
 - `apply`
@@ -55,22 +53,21 @@ Ask the user what command to run, or what additional information the user wants 
 - `verify`
 - `sync`
 - `onboard`
-- text input
+
+If the user input contains an `<action>`, build and run the command as defined above. For example, if the user input is "Please run ff to update the artifacts", you would execute `/opsx-ff`.
+
+If you are executing command `/opsx-ff` and after all artifacts are created/updated, call `/opsx-apply` in the next loop without asking the user for the command. Thus, all changes from the `ff` command will be applied immediately without asking the user for confirmation.
+
+If you are executing command `/opsx-apply`, implement changes till all tasks are completed. DO NOT STOP and  DO NOT ask anything till all tasks are completed. After all tasks are completed, move to the next step.
+
+If the user input does not contain any of the above actions, refine the artifacts based on the user input. For example, if the user input is "Please update the README file to include instructions for running the project", you would update all relevant artifacts in the current step to include updating the README.
+
+### D\[Ask open questions\]
+
+After the instructions are followed, identify all open questions defined in current artifacts. Ask the user these questions one by one and update the artifacts with the answers provided by the user. Open questions are defined by 'Open Questions' subtitle in the markdown files. Each question is defined as a bullet point under this subtitle.
 
 ### E\[Commit changes\]
 
 Descibe the changes being commited in the commit message. Check which `openspec` was used to make the changes and include that in the commit message. For example, if `openspec apply` was used, the commit message could be "apply: \<message which describes the changes\>".
 
 If the user answered with free-form text input in step A, use the last command used in the previous loops to define the commit message. For example, if the last command used was `ff`, the commit message could be "ff: \<message which describes the changes\>". If no command was used in previous loops, use "update: \<message which describes the changes\>" as the commit message.
-
-### C\[Call command\]
-
-Execute the command provided by the user. The command is defined as `/opsx-\<command\>`, where `\<command\>` is the user-provided command from step A. For example, if the user provides the command `ff`, you would execute `/opsx-ff`.
-
-If you are executing command `/opsx-ff` and after all artifacts are created/updated, call `/opsx-apply` in the next loop without asking the user for the command. Thus, all changes from the `ff` command will be applied immediately without asking the user for confirmation.
-
-If you are executing command `/opsx-apply`, implement changes till all tasks are completed. DO NOT STOP and  DO NOT ask anything till all tasks are completed. After all tasks are completed, move to the next step.
-
-### D\[Ask open questions\]
-
-After the command has finished executing, identify all open questions defined in current artifacts. Ask the user these questions one by one and update the artifacts with the answers provided by the user. Open questions are defined by 'Open Questions' subtitle in the markdown files. Each question is defined as a bullet point under this subtitle.
