+++ openspec/changes/fix-svg-preview-escape/.openspec.yaml
@@ -1,0 +1,2 @@
+schema: spec-driven
+created: 2026-04-14
